+++ app/components/sidebar.tsx
@@ -53,7 +53,7 @@ export function Sidebar({ homeName }: SidebarProps) {
               Groceries
             </Link>
             <Link
-              to="/recipe"
+              to="/recipes"
               prefetch="viewport"
               className={cn(
                 "flex items-center gap-3 rounded-lg px-3 py-2 text-muted-foreground transition-all hover:text-primary",
