+++ app/routes/_app.recipes_.new.tsx
@@ -1,0 +1,5 @@
+import { RecipeForm } from "@/components/recipe-form";
+
+export default function NewRecipePage() {
+  return <RecipeForm />;
+}
